Move common blocks into customblocks.common.link_blocks
