+++ network-api/networkapi/wagtailpages/pagemodels/customblocks/current_events_slider_list_block.py
@@ -1,7 +1,7 @@
 from wagtail.core import blocks
 from wagtail.images.blocks import ImageChooserBlock
 
-from .link_blocks import LabelledExternalLinkBlock, LabelledInternalLinkBlock
+from .common.link_blocks import LabelledExternalLinkBlock, LabelledInternalLinkBlock
 
 
 class CurrentEventBlock(blocks.StructBlock):
